requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e .
asgiref==3.5.0; python_version >= '3.7'
boto3==1.21.33
botocore==1.24.33; python_version >= '3.6'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
django-extensions==3.1.5
django-storages==1.12.3
django==3.2.12
djangorestframework-simplejwt==5.1.0
djangorestframework==3.13.1
drf-extra-fields==3.4.0
drf-yasg==1.20.0
google-api-core==2.7.1; python_version >= '3.6'
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth==2.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.56.0; python_version >= '3.6'
gunicorn==20.1.0
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.1; python_version >= '3.7'
jmespath==1.0.0; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
pillow==9.1.0
protobuf==3.20.0; python_version >= '3.7'
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyjwt==2.3.0; python_version >= '3.6'
pyparsing==3.0.7; python_version >= '3.1'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.6
pytz==2022.1
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
rsa==4.8; python_version >= '3.6'
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
s3transfer==0.5.2; python_version >= '3.6'
setuptools==62.0.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==6.0.0
